+++ Common/Integrations/GenericModConfigMenu/IGenericModConfigMenuApi.cs
@@ -1,0 +1,15 @@
+using System;
+using StardewModdingAPI;
+
+namespace Common.Integrations.GenericModConfigMenu
+{
+    /// <summary>The API provided by the Generic Mod Config Menu mod.</summary>
+    public interface IGenericModConfigMenuApi
+    {
+        void RegisterModConfig(IManifest mod, Action revertToDefault, Action saveToFile);
+        void RegisterLabel(IManifest mod, string labelName, string labelDesc);
+        void RegisterSimpleOption(IManifest mod, string optionName, string optionDesc, Func<bool> optionGet, Action<bool> optionSet);
+        void RegisterSimpleOption(IManifest mod, string optionName, string optionDesc, Func<SButton> optionGet, Action<SButton> optionSet);
+        void RegisterClampedOption(IManifest mod, string optionName, string optionDesc, Func<int> optionGet, Action<int> optionSet, int min, int max);
+    }
+}